small edits, opted for a different constants.py file too
