+++ firebase/firestore-py/lib/utils/__init__.py
@@ -2,4 +2,4 @@ from .logger import logger, log_value
 from .args import args
 from . import dir
 from . import constants
-constants.init()
+
